+++ .eslintrc
@@ -8,7 +8,7 @@
     "worker": true
   },
   "parserOptions": {
-    "ecmaVersion": 2018,
+    "ecmaVersion": 2020,
     "sourceType": "module"
   },
   "rules": {
@@ -34,6 +34,7 @@
     "no-undef": "warn",
     "no-case-declarations": "off",
     "array-callback-return": "off",
-    "camelcase": "warn"
+    "camelcase": "warn",
+    "allowImportExportEverywhere": 0
   }
 }
